Center ore veins and aura nodes on block in map

+++ src/main/java/com/sinthoras/visualprospecting/gui/model/locations/ILocationProvider.java
@@ -4,7 +4,7 @@ public interface ILocationProvider {
 
     int getDimensionId();
 
-    int getBlockX();
+    double getBlockX();
 
-    int getBlockZ();
+    double getBlockZ();
 }
